+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
-* Fixed `OnRequestStoppedCallback` exception handling when no root `Activity`
-  was created.
+* Fixed `OnRequestStoppedCallback` invocation when no root `Activity` was
+  created.
   ([#4307](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4307))
 
 ## 1.15.2